requirements.txt
discord.py
Pillow
cv2
numpy